# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BLE MVP API
  version: 0.1.0
  description: API v0.1 implementation contract. The product endpoints are not implemented yet.
servers:
  - url: /api/v1
security:
  - installationBearer: []
tags:
  - name: Installations
  - name: Me
  - name: Discovery
  - name: Chat
paths:
  /installations:
    post:
      tags: [Installations]
      summary: Register an installation
      operationId: createInstallation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallationRequest'
      responses:
        '201':
          description: A new installation and credential were created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationResponse'
        '200':
          description: The same request was retried within the 10-minute credential replay window.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationResponse'
        '409':
          description: IDEMPOTENCY_CONFLICT or IDEMPOTENCY_REPLAY_EXPIRED.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /me:
    get:
      tags: [Me]
      summary: Get the authenticated user
      operationId: getMe
      responses:
        '200':
          description: Current installation user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /me/profile:
    post:
      tags: [Me]
      summary: Create or replace the complete public profile
      operationId: putMyProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicProfile'
      responses:
        '200':
          description: Stored profile. Submitting identical values is a successful no-op.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /me/discovery:
    post:
      tags: [Me]
      summary: Set discovery state
      operationId: setDiscoveryState
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscoveryStateRequest'
      responses:
        '200':
          description: Current discovery state. Setting the same state is a successful no-op.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryStateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /me/push-token:
    post:
      tags: [Me]
      summary: Register this device for push notifications
      operationId: putPushToken
      description: >
        토큰은 기기가 아니라 **사용자** 단위로 보관된다. 같은 기기가 다른 사용자로
        다시 등록하면 토큰은 새 사용자에게 넘어가고 이전 사용자에게서는 사라진다.
        재설치나 기기 양도 뒤에 이전 사용자의 메시지가 계속 그 기기로 가는 일을
        막기 위한 것이다. 같은 값을 다시 보내는 것은 성공하는 no-op다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushTokenRequest'
      responses:
        '200':
          description: >
            등록됨. `push_enabled`가 false면 서버에 FCM이 설정되지 않은 상태이며,
            토큰은 그대로 보관되고 서버가 설정되면 발송이 시작된다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /discovery/identifiers:
    post:
      tags: [Discovery]
      summary: Get or atomically rotate the BLE temporary identifier
      operationId: issueDiscoveryIdentifier
      description: Returns the current identifier before four minutes and rotates it at or after four minutes. An identifier is valid for five minutes; the replaced value overlaps for at most one minute.
      responses:
        '200':
          description: Current or newly rotated identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryIdentifierResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: DISCOVERY_DISABLED or PROFILE_REQUIRED.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /discovery/observations:
    post:
      tags: [Discovery]
      summary: Report freshly scanned identifiers and return their public profiles
      operationId: reportDiscoveryObservations
      description: Invalid individual identifiers are ignored. Server receipt time becomes last_seen_at and grants first-conversation eligibility for ten minutes. No client observation timestamp is accepted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationRequest'
      responses:
        '200':
          description: Users resolved from valid observations. Each user occurs at most once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /discovery/recommendations/refresh:
    post:
      tags: [Discovery]
      summary: Refresh recommendations for previously observed users
      operationId: refreshDiscoveryRecommendations
      description: Reads current profiles and cached recommendations without renewing last_seen_at or conversation eligibility. Only users with an unexpired observation and discovery enabled are returned. Missing evaluations run in the background. Discovery disabled returns an empty list.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationRefreshRequest'
      responses:
        '200':
          description: Still eligible requested users, with current recommendations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: PROFILE_REQUIRED.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /conversations:
    get:
      tags: [Chat]
      summary: List my conversations
      operationId: listConversations
      description: Ordered by last message creation time descending.
      responses:
        '200':
          description: Conversations for the authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /messages:
    post:
      tags: [Chat]
      summary: Send a message
      operationId: createMessage
      description: A first message requires a valid sender-to-recipient observation from the last ten minutes and both users to have discovery enabled. Once a conversation exists, observation expiry, distance, and discovery state are not checked.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
      responses:
        '201':
          description: Message created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '200':
          description: An identical client_message_id, recipient, and text were retried; the original message is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: OBSERVATION_REQUIRED or DISCOVERY_DISABLED for a first message.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: RECIPIENT_NOT_FOUND.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: IDEMPOTENCY_CONFLICT or PROFILE_REQUIRED.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /conversations/{conversation_id}/messages:
    get:
      tags: [Chat]
      summary: Get messages after a conversation sequence
      operationId: listConversationMessages
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: after_seq
          in: query
          required: false
          description: Return messages whose server sequence is greater than this value.
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Messages ordered by seq ascending.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: CONVERSATION_NOT_FOUND for missing and non-participant conversations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  securitySchemes:
    installationBearer:
      type: http
      scheme: bearer
      bearerFormat: installation_credential
  responses:
    BadRequest:
      description: The JSON request could not be parsed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid installation credential.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request field validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: A required backing service is temporarily unavailable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    InstallationRequest:
      type: object
      additionalProperties: false
      required: [installation_request_id, platform]
      properties:
        installation_request_id:
          type: string
          format: uuid
          description: Client-generated UUID retained across retries of this registration.
        platform:
          type: string
          enum: [android]
    InstallationResponse:
      type: object
      additionalProperties: false
      required: [user_id, installation_credential, created_at]
      properties:
        user_id:
          $ref: '#/components/schemas/UserId'
        installation_credential:
          type: string
          minLength: 1
          description: Secret bearer credential. Replayable for only ten minutes through installation idempotency.
        created_at:
          $ref: '#/components/schemas/ServerTime'
    PublicProfile:
      type: object
      additionalProperties: false
      required: [nickname, self_description, connection_intent]
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 20
        self_description:
          type: string
          minLength: 1
          maxLength: 500
        connection_intent:
          type: string
          minLength: 1
          maxLength: 500
    MeResponse:
      type: object
      additionalProperties: false
      required: [user_id, profile, discovery_enabled]
      properties:
        user_id:
          $ref: '#/components/schemas/UserId'
        profile:
          oneOf:
            - $ref: '#/components/schemas/PublicProfile'
            - type: 'null'
        discovery_enabled:
          type: boolean
    DiscoveryStateRequest:
      type: object
      additionalProperties: false
      required: [enabled]
      properties:
        enabled:
          type: boolean
    DiscoveryStateResponse:
      type: object
      additionalProperties: false
      required: [discovery_enabled]
      properties:
        discovery_enabled:
          type: boolean
    DiscoveryIdentifierResponse:
      type: object
      additionalProperties: false
      required: [identifier, issued_at, refresh_after, expires_at]
      properties:
        identifier:
          $ref: '#/components/schemas/DiscoveryIdentifier'
        issued_at:
          $ref: '#/components/schemas/ServerTime'
        refresh_after:
          $ref: '#/components/schemas/ServerTime'
        expires_at:
          $ref: '#/components/schemas/ServerTime'
    ObservationRequest:
      type: object
      additionalProperties: false
      required: [identifiers]
      properties:
        identifiers:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/DiscoveryIdentifier'
    RecommendationRefreshRequest:
      type: object
      additionalProperties: false
      required: [user_ids]
      properties:
        user_ids:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/UserId'
    ObservationResponse:
      type: object
      additionalProperties: false
      required: [observed_users]
      properties:
        observed_users:
          type: array
          items:
            $ref: '#/components/schemas/ObservedUser'
    ObservedUser:
      type: object
      additionalProperties: false
      required: [user_id, profile, recommendation, last_seen_at, conversation_eligibility_expires_at]
      properties:
        user_id:
          $ref: '#/components/schemas/UserId'
        profile:
          $ref: '#/components/schemas/PublicProfile'
        recommendation:
          $ref: '#/components/schemas/Recommendation'
        last_seen_at:
          $ref: '#/components/schemas/ServerTime'
        conversation_eligibility_expires_at:
          $ref: '#/components/schemas/ServerTime'
    PushTokenRequest:
      type: object
      additionalProperties: false
      required: [token, platform]
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 4096
          description: FCM 등록 토큰. 형식은 Google 소유이므로 길이만 검사한다.
        platform:
          type: string
          enum: [android]
    PushTokenResponse:
      type: object
      additionalProperties: false
      required: [push_enabled]
      properties:
        push_enabled:
          type: boolean
          description: 서버가 실제로 푸시를 보낼 수 있는 상태인지. 등록 성공 여부와 별개다.
    Recommendation:
      type: object
      additionalProperties: false
      required: [status]
      description: >
        추천은 누가 근처에 있는지를 바꾸지 않는다. 평가가 없거나 실패해도 관측된
        사용자는 그대로 목록에 남으며, 이 객체는 그 사람을 어떻게 보여줄지만 정한다.
        점수·근거 인용·내부 버전·모델 정보는 공개하지 않는다.
      properties:
        status:
          type: string
          description: >
            `ready`는 평가가 끝나 이유가 있다는 뜻이고, `unscored`는 평가는 했지만
            특정할 만한 접점을 찾지 못했다는 뜻이다. 낮은 점수도 실패도 아니다.
            `pending`은 아직 평가 전이라 다음 관측에서 채워진다는 뜻이며,
            `failed`는 평가가 실패했다는 뜻이다. `unavailable`은 서버에 AI가
            설정되지 않은 상태다.
          enum: [ready, pending, unscored, failed, unavailable]
        rank:
          type: integer
          minimum: 0
          description: >
            이 응답에 담긴 사람들 사이의 표시 순서로, 0이 가장 위다. 응답마다 다시
            계산되므로 요청 사이에 안정적이지 않고, 다른 응답의 rank와 비교할 수
            없다. `unavailable`일 때는 정할 순서가 없으므로 생략된다.
        reason:
          type: string
          minLength: 1
          maxLength: 400
          description: >
            왜 이 사람인지 한두 문장. `status`가 `ready`일 때만 있고, 그 외에는
            생략된다(`null`이 아니다).
    CreateMessageRequest:
      type: object
      additionalProperties: false
      required: [recipient_id, client_message_id, text]
      properties:
        recipient_id:
          $ref: '#/components/schemas/UserId'
        client_message_id:
          type: string
          format: uuid
          description: Client-generated UUID retained across retries of this message.
        text:
          type: string
          minLength: 1
          maxLength: 2000
    Message:
      type: object
      additionalProperties: false
      required: [message_id, conversation_id, sender_id, recipient_id, seq, text, created_at]
      properties:
        message_id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        sender_id:
          $ref: '#/components/schemas/UserId'
        recipient_id:
          $ref: '#/components/schemas/UserId'
        seq:
          type: integer
          minimum: 1
          description: Server-issued monotonically increasing sequence within the conversation.
        text:
          type: string
          minLength: 1
          maxLength: 2000
        created_at:
          $ref: '#/components/schemas/ServerTime'
    ConversationParticipant:
      type: object
      additionalProperties: false
      required: [user_id, profile]
      properties:
        user_id:
          $ref: '#/components/schemas/UserId'
        profile:
          $ref: '#/components/schemas/PublicProfile'
    ConversationSummary:
      type: object
      additionalProperties: false
      required: [conversation_id, participant, last_message]
      properties:
        conversation_id:
          type: string
          format: uuid
        participant:
          $ref: '#/components/schemas/ConversationParticipant'
        last_message:
          $ref: '#/components/schemas/Message'
    ConversationListResponse:
      type: object
      additionalProperties: false
      required: [conversations]
      properties:
        conversations:
          type: array
          items:
            $ref: '#/components/schemas/ConversationSummary'
    MessagePage:
      type: object
      additionalProperties: false
      required: [messages, next_after_seq, has_more]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        next_after_seq:
          oneOf:
            - type: integer
              minimum: 1
            - type: 'null'
        has_more:
          type: boolean
    UserId:
      type: string
      format: uuid
      description: Public identifier; never an authentication credential.
    DiscoveryIdentifier:
      type: string
      minLength: 22
      maxLength: 22
      pattern: '^[A-Za-z0-9_-]{22}$'
      description: Unpadded base64url encoding of 128 random bits.
    ServerTime:
      type: string
      format: date-time
      description: UTC RFC 3339 timestamp issued by the server.
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details]
          properties:
            code:
              type: string
              minLength: 1
            message:
              type: string
              minLength: 1
            details:
              type: object
              additionalProperties: true